# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DirectorInstruction": {
        "additionalProperties": false,
        "properties": {
          "applies_from_checkpoint_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Applies From Checkpoint Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "instruction_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Instruction Id",
            "type": "string"
          },
          "text": {
            "maxLength": 65536,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "instruction_id",
          "text",
          "created_at",
          "applies_from_checkpoint_id"
        ],
        "title": "DirectorInstruction",
        "type": "object"
      },
      "DirectorInstructionRequest": {
        "additionalProperties": false,
        "properties": {
          "checkpoint_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Checkpoint Id",
            "type": "string"
          },
          "text": {
            "maxLength": 65536,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "checkpoint_id",
          "text"
        ],
        "title": "DirectorInstructionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WikiBranchView": {
        "additionalProperties": false,
        "properties": {
          "branch_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Branch Id",
            "type": "string"
          },
          "checkpoint_id": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Id"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/WikiPageSummary"
            },
            "title": "Pages",
            "type": "array"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          },
          "updated_at_step": {
            "minimum": 0.0,
            "title": "Updated At Step",
            "type": "integer"
          }
        },
        "required": [
          "branch_id",
          "updated_at_step",
          "stale",
          "pages"
        ],
        "title": "WikiBranchView",
        "type": "object"
      },
      "WikiLintIssue": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Code",
            "type": "string"
          },
          "message": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "path": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "severity": {
            "$ref": "#/components/schemas/WikiLintSeverity"
          }
        },
        "required": [
          "code",
          "message",
          "severity"
        ],
        "title": "WikiLintIssue",
        "type": "object"
      },
      "WikiLintResult": {
        "additionalProperties": false,
        "properties": {
          "branch_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Branch Id",
            "type": "string"
          },
          "issues": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/WikiLintIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          }
        },
        "required": [
          "branch_id",
          "passed"
        ],
        "title": "WikiLintResult",
        "type": "object"
      },
      "WikiLintSeverity": {
        "enum": [
          "warning",
          "error"
        ],
        "title": "WikiLintSeverity",
        "type": "string"
      },
...
      "WikiPage": {
        "additionalProperties": false,
        "properties": {
          "branch_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Branch Id",
            "type": "string"
          },
          "checkpoint_id": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Id"
          },
          "confidence": {
            "default": 1.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "content": {
            "maxLength": 131072,
            "title": "Content",
            "type": "string"
          },
          "path": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          },
          "source_ids": {
            "default": [],
            "items": {
              "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          },
          "stale": {
            "default": false,
            "title": "Stale",
            "type": "boolean"
          },
          "subject_id": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          },
          "updated_at_step": {
            "minimum": 0.0,
            "title": "Updated At Step",
            "type": "integer"
          }
        },
        "required": [
          "branch_id",
          "path",
          "updated_at_step",
          "content"
        ],
        "title": "WikiPage",
        "type": "object"
      },
      "WikiPageSummary": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "source_ids": {
            "default": [],
            "items": {
              "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          },
          "subject_id": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at_step": {
            "minimum": 0.0,
            "title": "Updated At Step",
            "type": "integer"
          }
        },
        "required": [
          "path",
          "title",
          "updated_at_step",
          "confidence"
        ],
        "title": "WikiPageSummary",
        "type": "object"
      },
      "WikiRebuildRequest": {
        "additionalProperties": false,
        "properties": {
          "checkpoint_id": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Id"
          }
        },
        "title": "WikiRebuildRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionToken": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Local story-domain sidecar API",
    "title": "AI Story Evolution Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/projects/{project_id}/branches/{branch_id}/director-instructions": {
      "get": {
        "operationId": "list_director_instructions_projects__project_id__branches__branch_id__director_instructions_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DirectorInstruction"
                  },
                  "title": "Response List Director Instructions Projects  Project Id  Branches  Branch Id  Director Instructions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "List Director Instructions",
        "tags": [
          "wiki"
        ]
      },
      "post": {
        "operationId": "add_director_instruction_projects__project_id__branches__branch_id__director_instructions_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectorInstructionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectorInstruction"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Add Director Instruction",
        "tags": [
          "wiki"
        ]
      }
    },
...
    "/projects/{project_id}/branches/{branch_id}/wiki": {
      "get": {
        "operationId": "get_wiki_projects__project_id__branches__branch_id__wiki_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WikiBranchView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Get Wiki",
        "tags": [
          "wiki"
        ]
      }
    },
    "/projects/{project_id}/branches/{branch_id}/wiki/lint": {
      "get": {
        "operationId": "lint_wiki_projects__project_id__branches__branch_id__wiki_lint_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WikiLintResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Lint Wiki",
        "tags": [
          "wiki"
        ]
      }
    },
    "/projects/{project_id}/branches/{branch_id}/wiki/page": {
      "get": {
        "operationId": "get_wiki_page_projects__project_id__branches__branch_id__wiki_page_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "maxLength": 512,
              "minLength": 1,
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WikiPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Get Wiki Page",
        "tags": [
          "wiki"
        ]
      }
    },
    "/projects/{project_id}/branches/{branch_id}/wiki/rebuild": {
      "post": {
        "operationId": "rebuild_wiki_projects__project_id__branches__branch_id__wiki_rebuild_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WikiRebuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WikiBranchView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Rebuild Wiki",
        "tags": [
          "wiki"
        ]
      }
    },
...
  }
}